move: 패키지 이동
- /entity -> /entity/id
